requirements.txt
audioread==3.0.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
decorator==4.4.2
facenet-pytorch==2.5.2
ffmpeg==1.4
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
librosa==0.8.1
llvmlite==0.39.1
moviepy==1.0.3
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
platformdirs==3.0.0
pooch==1.7.0
proglog==0.1.10
pycparser==2.21
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.1
scipy==1.10.1
soundfile==0.12.1
threadpoolctl==3.1.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
wincertstore==0.2
